requirements.txt
#requirements.txt
snowflake-connector-python
snowflake-snowpark-python
streamlit==1.31.1